renamed social to friends
