+++ student_app/lib/features/friends/friends_profile_popup.dart
