+++ student_app/lib/features/friends/friends_user_notification_popup.dart
